delete form added

+++ app/api/comment_routes.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, request, render_template
 from flask_login import login_required, current_user
 from app.models import db, User, Comment
 from app.forms.comment_form import CommentForm
+from app.forms.delete_form import DeleteForm
 from app.api.project_routes import validation_errors_to_error_messages
 
 comment_routes = Blueprint('comments', __name__)
@@ -28,13 +29,17 @@ def update_comment(id):
 
 
 @comment_routes.route('/<int:id>', methods=["DELETE"])
-# tmp comment out
-# @login_required
+@login_required
 def delete_comment(id):
-    comment = Comment.query.get(id)
-    if comment:
-        db.session.delete(comment)
-        db.session.commit()
-        return {"message": f"Comment {id} was deleted successfully"}
-    else:
-        return {'errors': ["Comment not found"]}, 404
+    form = DeleteForm()
+    form['csrf_token'].data = request.cookies['csrf_token']
+
+    if form.validate_on_submit():
+        comment = Comment.query.get(id)
+        if comment:
+            db.session.delete(comment)
+            db.session.commit()
+            return {"message": f"Comment {id} was deleted successfully"}
+        else:
+            return {'errors': ["Comment not found"]}, 404
+    return {'errors': ['An error has occurred. Please try again.']}, 401
